Move PyConvertible trait to `py.rs`
